+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ### 🐛 Bug fixes
 
-- Fix the following Bmad lattice conversion bugs: skips parsing lines with overlay or group definitions, wildcard element references, scientific notation, +/- before unary operators, maps type/alias fields to element metadata to reduce the number of warnings (see #663) (@roussel-ryan)
+- Fix Bmad lattice conversion issues with overlay and group definitions, wildcard element references, scientific notation in expressions, unary signs, and mapping of type and alias fields to element metadata (see #663) (@roussel-ryan, @jank324)
 
 ### 🐆 Other